pyproject.toml
[project]
name = "local-media-viewer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anyio>=4.11.0",
    "fastapi[standard]>=0.118.0",
    "jinja2>=3.1.6",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.11.0",
    "pytz>=2025.2",
    "pyyaml>=6.0.3",
]
